Renamed plugin run classes to main
