+++ bootstrap/bukkit/pom.xml
@@ -40,7 +40,7 @@
                 <configuration>
                     <archive>
                         <manifestEntries>
-                            <Main-Class>org.geysermc.platform.bukkit.GeyserBukkitRunnable</Main-Class>
+                            <Main-Class>org.geysermc.platform.bukkit.GeyserBukkitMain</Main-Class>
                         </manifestEntries>
                     </archive>
                 </configuration>
