+++ bootstrap/sponge/pom.xml
@@ -40,7 +40,7 @@
                 <configuration>
                     <archive>
                         <manifestEntries>
-                            <Main-Class>org.geysermc.platform.sponge.GeyserSpongeRunnable</Main-Class>
+                            <Main-Class>org.geysermc.platform.sponge.GeyserSpongeMain</Main-Class>
                         </manifestEntries>
                     </archive>
                 </configuration>
